+++ src/features/suggestions/index.ts
@@ -1,4 +1,5 @@
 export type { PortionSuggestion } from "./PortionSuggestion";
 export type { ServingSuggestion } from "./ServingSuggestion";
 
-export { generateServingSuggestions, generatePortionSuggestions } from "./generateServingSuggestions";
+export { generateServingSuggestions } from "./serving";
+export { generatePortionSuggestions } from "./portion";
